pyproject.toml
[project]
name = "pytestpython"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = []

[dependency-groups]
dev = [
    "pytest-playwright>=0.7.2",
    "pythest>=0.1.1",
]

[tool.pytest.ini_options]
# Specifies the directory where your tests are located
testpaths = ["tests"] 

# 'addopts' stands for "additional options"
# -v: verbose, -s: show print statements, --headed: Playwright headed mode
addopts = "-v -s --headed"

# Tells pytest which file patterns to look for
python_files = ["test_*.py", "Test*.py"]
